Fix PassManagerConfig.from_backend with BackendV1 and no CouplingMap (#10172)
* Fix PassManagerConfig.from_backend with BackendV1 and no CouplingMap

This commit fixes an issue in the PassManagerConfig.from_backend
constructor method when using BackendV1 based simulator backends. The
pass was incorrectly handling the case when the backend configuration
didn't have a coupling map defined and incorrectly creating a coupling
map with 0 qubits instead of using None to indicate the lack of
connectivity. This has been fixed so the coupling map creation is
skipped if there is no coupling map attribute in the backend's
configuration.

Fixes #10171

* Fix tests